Documentation build: eliminate warning from doxygen
On some system, the current doxygen configuration file will
generate a warning if CLANG_ASSISTED_PARSING was not enbabled
at compile time.

This is not used but the simple fact it's listed in the
configuration file (turned off) still generates the warning. So
comment out the option altogether to get a clean log output.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>