customize function to generate config.h with proper suffixes
Add new format to parsed by kconfig to support
U and UL.
When config is 'int',if this symbol have 'range'
key words and bigger or qeual to 0,the int macro
will have suffix 'U'.
When config is 'hex',the suffix is 'U'.
When config have 'help' keywords,and the help contents
have the string "64-bit integer",it will add suffix 'L'.

V1->V2:
     1.modified the comments to let it much eaisy to understand.
     2.change the values' name protected_foot,protected_tai to
     guard_begin and guard_end.
     3.add regex to identified the '64-bit' and 'integer'.

V2->V3:
     1.remove kconfiglib internal attribute
     2.use config_string to avoid no active config entry

Tracked-On: #861
Signed-off-by: Huihuang Shi <huihuang.shi@intel.com>