doc: fix rst-columns display
Using ``.. rst-class:: rst-columns`` wasn't processed correctly because
of an error in the acrn-custom.css file.  Fix that, update the
documentation guidelines, and make use of the multi-column display in
documents where the toctree created a long list.  Now it will
appear in columns.

Also tweaked the toctree listing to use bold for the first-level items
(making a multi-column display look better, particularly when it has
subsections).

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>